+++ nodeconductor/openstack/backend.py
@@ -220,7 +220,7 @@ class OpenStackBackend(ServiceBackend):
         raise AttributeError(
             "'%s' object has no attribute '%s'" % (self.__class__.__name__, name))
 
-    def ping(self):
+    def ping(self, raise_exception=False):
         # Session validation occurs on class creation so assume it's active
         # TODO: Consider validating session depending on tenant permissions
         return True
@@ -477,7 +477,7 @@ class OpenStackBackend(ServiceBackend):
                 time.sleep(poll_interval)
 
             return False
-        except cinder_exceptions.NotFound:
+        except (cinder_exceptions.NotFound, keystone_exceptions.NotFound):
             return True
 
     def _wait_for_instance_deletion(self, backend_instance_id, retries=90, poll_interval=3):
@@ -1361,7 +1361,7 @@ class OpenStackBackend(ServiceBackend):
                 logger.warning('Not extending volume %s: desired size %d MiB is less then current size %d MiB',
                                volume_id, new_core_size, old_core_size)
                 event_logger.openstack_volume.error(
-                    "Virtual machine {instance_name} disk extension has failed "
+                    "Virtual machine {resource_name} disk extension has failed "
                     "due to new size being less than old size.",
                     event_type='resource_volume_extension_failed',
                     event_context={'resource': instance}
@@ -1387,7 +1387,7 @@ class OpenStackBackend(ServiceBackend):
                     volume.id,
                 )
                 event_logger.openstack_volume.error(
-                    "Virtual machine {instance_name} disk extension has failed due to quota limits.",
+                    "Virtual machine {resource_name} disk extension has failed due to quota limits.",
                     event_type='resource_volume_extension_failed',
                     event_context={'resource': instance},
                 )
@@ -1417,7 +1417,7 @@ class OpenStackBackend(ServiceBackend):
         else:
             logger.info('Successfully extended disk of an instance %s', instance.uuid)
             event_logger.openstack_volume.info(
-                "Virtual machine {instance_name} disk has been extended to {volume_size} GB.",
+                "Virtual machine {resource_name} disk has been extended to {volume_size} GB.",
                 event_type='resource_volume_extension_succeeded',
                 event_context={'resource': instance, 'volume_size': new_core_size_gib},
             )
@@ -1862,14 +1862,14 @@ class OpenStackBackend(ServiceBackend):
             if not self._wait_for_instance_status(instance.backend_id, nova, 'ACTIVE'):
                 logger.error('Failed to start instance %s', instance.uuid)
                 event_logger.resource.error(
-                    'Virtual machine {instance_name} start has failed.',
+                    'Virtual machine {resource_name} start has failed.',
                     event_type='resource_start_failed',
                     event_context={'resource': instance})
                 raise OpenStackBackendError('Timed out waiting for instance %s to start' % instance.uuid)
         except nova_exceptions.ClientException as e:
             logger.exception('Failed to start instance %s', instance.uuid)
             event_logger.resource.error(
-                'Virtual machine {instance_name} start has failed.',
+                'Virtual machine {resource_name} start has failed.',
                 event_type='resource_start_failed',
                 event_context={'resource': instance})
             six.reraise(OpenStackBackendError, e)
@@ -1879,7 +1879,7 @@ class OpenStackBackend(ServiceBackend):
 
             logger.info('Successfully started instance %s', instance.uuid)
             event_logger.resource.info(
-                'Virtual machine {instance_name} has been started.',
+                'Virtual machine {resource_name} has been started.',
                 event_type='resource_start_succeeded',
                 event_context={'resource': instance})
 
@@ -1899,14 +1899,14 @@ class OpenStackBackend(ServiceBackend):
             if not self._wait_for_instance_status(instance.backend_id, nova, 'SHUTOFF'):
                 logger.error('Failed to stop instance %s', instance.uuid)
                 event_logger.resource.error(
-                    'Virtual machine {instance_name} stop has failed.',
+                    'Virtual machine {resource_name} stop has failed.',
                     event_type='resource_stop_failed',
                     event_context={'resource': instance})
                 raise OpenStackBackendError('Timed out waiting for instance %s to stop' % instance.uuid)
         except nova_exceptions.ClientException as e:
             logger.exception('Failed to stop instance %s', instance.uuid)
             event_logger.resource.error(
-                'Virtual machine {instance_name} stop has failed.',
+                'Virtual machine {resource_name} stop has failed.',
                 event_type='resource_stop_failed',
                 event_context={'resource': instance})
             six.reraise(OpenStackBackendError, e)
@@ -1915,7 +1915,7 @@ class OpenStackBackend(ServiceBackend):
             instance.save(update_fields=['start_time'])
             logger.info('Successfully stopped instance %s', instance.uuid)
             event_logger.resource.info(
-                'Virtual machine {instance_name} has been stopped.',
+                'Virtual machine {resource_name} has been stopped.',
                 event_type='resource_stop_succeeded',
                 event_context={'resource': instance})
 
@@ -1928,21 +1928,21 @@ class OpenStackBackend(ServiceBackend):
             if not self._wait_for_instance_status(instance.backend_id, nova, 'ACTIVE', retries=80):
                 logger.error('Failed to restart instance %s', instance.uuid)
                 event_logger.resource.error(
-                    'Virtual machine {instance_name} restart has failed.',
+                    'Virtual machine {resource_name} restart has failed.',
                     event_type='resource_restart_failed',
                     event_context={'resource': instance})
                 raise OpenStackBackendError('Timed out waiting for instance %s to restart' % instance.uuid)
         except nova_exceptions.ClientException as e:
             logger.exception('Failed to restart instance %s', instance.uuid)
             event_logger.resource.error(
-                'Virtual machine {instance_name} restart has failed.',
+                'Virtual machine {resource_name} restart has failed.',
                 event_type='resource_restart_failed',
                 event_context={'resource': instance})
             six.reraise(OpenStackBackendError, e)
         else:
             logger.info('Successfully restarted instance %s', instance.uuid)
             event_logger.resource.info(
-                'Virtual machine {instance_name} has been restarted.',
+                'Virtual machine {resource_name} has been restarted.',
                 event_type='resource_restart_succeeded',
                 event_context={'resource': instance})
 
@@ -1955,7 +1955,7 @@ class OpenStackBackend(ServiceBackend):
             if not self._wait_for_instance_deletion(instance.backend_id):
                 logger.info('Failed to delete instance %s', instance.uuid)
                 event_logger.resource.error(
-                    'Virtual machine {instance_name} deletion has failed.',
+                    'Virtual machine {resource_name} deletion has failed.',
                     event_type='resource_deletion_failed',
                     event_context={'resource': instance})
                 raise OpenStackBackendError('Timed out waiting for instance %s to get deleted' % instance.uuid)
@@ -1963,14 +1963,14 @@ class OpenStackBackend(ServiceBackend):
         except nova_exceptions.ClientException as e:
             logger.info('Failed to delete instance %s', instance.uuid)
             event_logger.resource.error(
-                'Virtual machine {instance_name} deletion has failed.',
+                'Virtual machine {resource_name} deletion has failed.',
                 event_type='resource_deletion_failed',
                 event_context={'resource': instance})
             six.reraise(OpenStackBackendError, e)
         else:
             logger.info('Successfully deleted instance %s', instance.uuid)
             event_logger.resource.info(
-                'Virtual machine {instance_name} has been deleted.',
+                'Virtual machine {resource_name} has been deleted.',
                 event_type='resource_deletion_succeeded',
                 event_context={'resource': instance})
 
@@ -2078,3 +2078,25 @@ class OpenStackBackend(ServiceBackend):
                 six.reraise(OpenStackBackendError, e)
         else:
             logger.warning('Cannot update tenant name for link %s without tenant ID', service_project_link)
+
+    def create_snapshot(self, volume_id, cinder):
+        """
+        Create snapshot from volume
+
+        :param: volume id
+        :type volume_id: str
+        :returns: snapshot id
+        :rtype: str
+        """
+        snapshot = cinder.volume_snapshots.create(
+            volume_id, force=True, display_name='snapshot_from_volume_%s' % volume_id)
+
+        logger.debug('About to create temporary snapshot %s' % snapshot.id)
+
+        if not self._wait_for_snapshot_status(snapshot.id, cinder, 'available', 'error'):
+            logger.error('Timed out creating snapshot for volume %s', volume_id)
+            raise OpenStackBackendError()
+
+        logger.info('Successfully created snapshot %s for volume %s', snapshot.id, volume_id)
+
+        return snapshot